requirements.txt
async-timeout==4.0.2
click==8.1.3
commonmark==0.9.1
croniter==1.3.8
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
hiredis==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
nanoid==2.0.0
orjson==3.8.5
pydantic==1.10.4
Pygments==2.14.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.1
redis==4.4.2
rich==12.6.0
six==1.16.0
SQLAlchemy==1.4.46
typing_extensions==4.4.0
Werkzeug==2.2.2
